git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@68950 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/aimsdata/sparsematrix/sparseMatrix.cc
@@ -916,7 +916,7 @@ std::vector<double> aims::SparseMatrix::getRow( int32_t i ) const
     std::vector<double> row( getSize2(), 0. );
     boost_sparse_matrix::const_iterator1
       ir = _matrix.find1( 0, i, 0 );
-    if( ir != _matrix.end1() && ir.index1() == i )
+    if( ir != _matrix.end1() && ir.index1() == static_cast<size_t>(i) )
     {
       boost_sparse_matrix::const_iterator2
         ic, ec = ir.end();
